[#156] fix: 긴 글에서만 사용되는 유틸 함수 분리
