+++ client/src/components/practice/long/Content/index.tsx
@@ -1,10 +1,10 @@
 import { Flex } from '@chakra-ui/react';
 
 import type { LongTypingDetail } from '@/types/typing';
-import { slicedContentAndStrings } from '@/utils/typing';
 
 import TypingLine from '../common/TypingLine';
 import PracticeLongLayout from '../Layout';
+import { slicedContentAndStrings } from '../utils';
 
 export default function LongContent({ content, currentPage, language, title, totalPage, typingId }: LongTypingDetail) {
   return (
